feat: add macro for generating schema adapter methods in FileSource implementations
